### 2024-08-20 d0353bd
Moved contributing guidelines to a separate file. Updated DOIs in paper.bib.
