+++ paper/paper.bib
@@ -4,7 +4,8 @@
   author={Aspuru-Guzik, Al{\'a}n and Persson, Kristin},
   journal={Mission Innovation},
   year={2018},
-  publisher={Canadian Institute for Advanced Research}
+  publisher={Canadian Institute for Advanced Research},
+  uri={http://nrs.harvard.edu/urn-3:HUL.InstRepos:35164974}
 }
 
 @article{maier2007combinatorial,
@@ -15,7 +16,8 @@
   number={32},
   pages={6016--6067},
   year={2007},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/anie.200603675}
 }
 
 @article{wolf2023accelerated,
@@ -26,7 +28,8 @@
   number={12},
   pages={2612--2624},
   year={2023},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi={10.1039/D3MA00136A}
 }
 
 @unpublished{wolf2024unpublished,
@@ -42,7 +45,8 @@
   volume={117},
   pages={176},
   year={2012},
-  publisher={National Institute of Standards and Technology}
+  publisher={National Institute of Standards and Technology},
+  doi={10.6028/jres.117.010}
 }
 
 @article{weber2021pymodaq,
@@ -52,14 +56,21 @@
   volume={92},
   number={4},
   year={2021},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi={10.1063/5.0032116}
 }
 
 @article{fuchs2024nomad,
-  title={NOMAD CAMELS: Configurable Application for Measurements, Experiments and Laboratory Systems},
-  author={Fuchs, Alexander D and Lehmeyer, Johannes AF and Junkes, Heinz and Weber, Heiko B and Krieger, Michael},
-  journal={arXiv preprint arXiv:2402.07548},
-  year={2024}
+  doi = {10.21105/joss.06371},
+  url = {https://doi.org/10.21105/joss.06371},
+  year = {2024},
+  publisher = {The Open Journal},
+  volume = {9},
+  number = {95},
+  pages = {6371},
+  author = {Alexander D. Fuchs and Johannes A. f. Lehmeyer and Heinz Junkes and Heiko B. Weber and Michael Krieger},
+  title = {NOMAD CAMELS: Configurable Application for Measurements, Experiments and Laboratory Systems},
+  journal = {Journal of Open Source Software}
 }
 
 @article{allan2019bluesky,
@@ -70,5 +81,6 @@
   number={3},
   pages={19--22},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/08940886.2019.1608121}
 }
