### 2024-08-20 #7122 @mxwalbert
> @mxwalbert, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
> * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
> * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

Thank you!

- [x] I updated the DOIs in the paper.bib,
- [x] and moved the contributing guidelines from the README to a separate file CONTRIBUTING.md in the repository root.
