### 2024-08-23 #7122 @RMeli
Hi @samaygarg 👋

Would you be interested in reviewing this submission for JOSS (Journal of Open Source Software)?

You can find more information about the submission at the top of this Github issue #7122.

Thank you in advance!
