### 2024-10-10 #7291 @enricgrau
@mxwalbert Thank you for your submission to JOSS. I went through the paper and repository and I have concerns about the scholarly effort and necessity of the library. Here I list one by one:
1)	Need and Novelty: the manuscript fails to show the problems in the field what the library attempts to solve. At the same time, it is not clear at all what the library brings to the table. The manuscript should clearly respond to the questions: What gaps exist in the field that this library fills? What software currently exists, if any, and why can’t they tackle those issue? How does this library compare to and trumps over those?
2)	Functionality not mentioned: you make a good attempt on briefly mentioning the importance of accelerating materials research and combinatorial analysis. However, there is no mention on exactly how or what the library helps you do. The manuscript includes vague and generic phrases such as “simplify the process of…” and “…provides researchers with a with flexible and modular framework…”. The authors should aim to explain more specifically what these advantages are so the reader can more effectively understand if it is useful to them.
3)	Insufficient scholarly effort: It is not clear to me that the library gathers enough work to be published. Even though the repository has existed for over a year, it only shows a couple of periods with activity and just a handful of commits. The fact that is has been used in previous work does not mean that there was enough effort behind the library. In this same regard, the work is not published or available as pre-print, so there is no way to even verify this.
4)	Lis of authors: the main author has been the main contributor without a doubt, with just a couple of contributed lines by who I believe to be the second author. The other three authors have not committed any code, so it is unclear to me why they are listed as authors. I recommend using CRediT to assign the corresponding roles to each author. You can do this in the manuscript as a separate section at the end.

Because of these concerns I have not checked the library or tried the code at all. I’ll proceed to do so after the author successfully addresses these points. Please note that these points are trying to be constructive, you may have a good piece of software here but it is just no reflecting that. I’m happy to clarify and discuss all the above and help to address some of the points mentioned.
