### 2024-10-14 #7291 @mxwalbert
@enricgrau Thank you for pointing out your concerns about our submission. I want to address them one by one:

> 1. Need and Novelty: the manuscript fails to show the problems in the field what the library attempts to solve. At the same time, it is not clear at all what the library brings to the table. The manuscript should clearly respond to the questions: What gaps exist in the field that this library fills? What software currently exists, if any, and why can’t they tackle those issue? How does this library compare to and trumps over those?

In the current manuscript, we attempt to answer these questions in the _Statement of need_ section. We mention the purpose "_(...)simplify the process of setting up and executing combinatorial voltaic measurements._" and briefly compare the software against other existing tools in the second paragraph. Could you please concretize what exactly raised your concern?

> 2. Functionality not mentioned: you make a good attempt on briefly mentioning the importance of accelerating materials research and combinatorial analysis. However, there is no mention on exactly how or what the library helps you do. The manuscript includes vague and generic phrases such as “simplify the process of…” and “…provides researchers with a with flexible and modular framework…”. The authors should aim to explain more specifically what these advantages are so the reader can more effectively understand if it is useful to them.

We try to summarize the high-level functionality and advantages of the software in the first paragraph of the _Statement of need_ section. However, we apparently fail to do so. The "vague and generic phrases" you mention are intentional because they specifically point out the advantages. We designed the software to be as abstract/generic as possible in order to make it usable for a large audience. Researchers should be able to use their measurement equipment with our software to run combinatorial measurements. Certainly, they still have to implement an API for their specific hardware to interface it with our software, but we provide an extensive guide in the documentation to do so. Then, the advantages will come into play because it's easy to setup and run experiments, store the data, collect relevant metadata and analyse the results. Furthermore, the implemented components can be reused which enables you to run a [``Measurement``](https://cohesivm.readthedocs.io/en/latest/reference/measurements.html#cohesivm.measurements.Measurement) using different [``Device``](https://cohesivm.readthedocs.io/en/latest/reference/devices.html#cohesivm.devices.Device)s. We intentionally left this level of detail out of the manuscript because it is written in the repository's readme. Do you think describing this briefly would resolve your concern?

> 3. Insufficient scholarly effort: It is not clear to me that the library gathers enough work to be published. Even though the repository has existed for over a year, it only shows a couple of periods with activity and just a handful of commits. The fact that is has been used in previous work does not mean that there was enough effort behind the library. In this same regard, the work is not published or available as pre-print, so there is no way to even verify this.

For this submission, the scholary effort should not be measured based on the number of commits. We do not have a background in software development and most of the work was done locally before even considering a publication. I think the effort behind this project becomes evident easily by looking at the source code and documentation. As for the cited previous work, we are currently drafting the paper but some effort can be verified by the checking the hardware documentation in the repository: [MA8X8](https://github.com/mxwalbert/cohesivm/tree/main/hardware/ma8x8).

> 4. Lis of authors: the main author has been the main contributor without a doubt, with just a couple of contributed lines by who I believe to be the second author. The other three authors have not committed any code, so it is unclear to me why they are listed as authors. I recommend using CRediT to assign the corresponding roles to each author. You can do this in the manuscript as a separate section at the end.

The same reason from above applies here. But certainly, we will include a CRediT section in the revision to clarify the contributions.

Please let me know, if possible, how we should address concern 1 and if concerns 2-4 can be resolved as suggested in my comments.
