### 2024-10-14 #7291 @enricgrau
@mxwalbert
For 1) and 2): I went over the manuscript again and I'm still not sure what the library does and why it exists. For instance, the first phrase in the _Statement of need_ declares that it "... aims to simplify the process of setting up and executing combinatorial voltaic measurements." Why does this process need to be simplified? How does it make it more simple? Try to follow that sentence with "... by doing XYZ, which makes it simpler." or something along those ways. Another example is when you claim that it is _simpler to use_. How? Why? What makes other options complicated to implement? I'd recommend focusing on its specific functionality and being more specific overall. By doing that it will reflect its simplicity and other claims wihtout the need on emphazising so much on buzzwords. Maybe including something like _Key features_ in the manuscript could help. Is COHESIVM for data colection? Data processing? Visalization? System control? All of the above? I'm genuinely not sure after reading the manuscript. In a few words, my issue is that I don't think the manuscript projects to be sufficiently useful or likely to be cited or used.

For 3): As per the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), the number of commits is a factor we can use to judge scholarly effort. Any chance you can upload a draft to arXiv? That'd be enough to support the claim of the library being used before. Using the docs to support that is too much of self-sustained argument in my opinion (ie citing a paper in the same paper.)
