### 2024-10-14 #7291 @mxwalbert
@enricgrau
Thank you for clarification. I will follow your suggestions to phrase a more specific description.

I agree that the number of commits _can_ be used but it's not a requirement in my understanding. This is why I referred to the repository and documentation to base the judgement on them. Unfortunately, the draft is not ready yet, as we plan to finish it until end of November. However, the paper will be a completely separate contribution and only briefly mention the use of this software. So everything within the repository is open for judging the effort of this submission since it is not and will not be published elsewhere.
