### 2024-10-29 mxwalbert/cohesivm#1 @ericfell
[JOSS Review] - testing fails due to incorrect dev instructions in README
Dear authors,

Given your current configuration of `pyproject.toml`, the clone and install process for development version fails when running `pytest test/` because `pyvisa` isn't a required install.

The remedy is to change the README instruction #3 from `pip install .` to `pip install .[dev]`.

This issue is related to openjournals/joss-reviews#7291
