### 2024-10-29 mxwalbert/cohesivm#2 @ericfell
[JOSS Review] - Community guidelines
Dear authors,

Thank you for including a `CONTRIBUTING.md`.
Can you also add a **Contributing** section somewhere in `docs` (so that it appears on readthedocs) to improve the visibility for third parties looking to contribute? This could paraphrase `CONTRIBUTING.md` or just provide a link to it. Often this appears in a FAQ page in readthedocs.

This issue is related to openjournals/joss-reviews#7291
