### 2024-10-29 mxwalbert/cohesivm#3 @ericfell
[JOSS Review] - Suggestions for configuration documentation
Dear authors,

In the `README`'s `Configuration` section, can you provide further documentation that covers the situation where a user wants to configure multiple devices and use them together (i.e. multiple COM ports being used)?
A short code example where multiple devices are configured in a single `config.ini` file could go just below the single-device example shown at https://cohesivm.readthedocs.io/en/latest/getting_started/configuration.html

Configuring new devices will likely be the situation where significant bugs/troubleshooting could occur, and thus significant documentation/examples will benefit all users.

This issue is related to openjournals/joss-reviews#7291
