### 2024-10-29 mxwalbert/cohesivm#4 @ericfell
[JOSS Review] - Suggestion for tutorial documentation
Dear authors,

The five _implement.../...workflow_ tutorials provided are well documented, but I feel that the tutorial section could benefit from a few initial sentences of high-level description of an example of a device a user would want to connect, and the subsequent experiment they'd perform.
This could be described with mentions to `Channel/Device/Interface/Experiment/Measurement` modules and subsequently, `Analysis`. This high-level-example description could be included in the README and the JOSS paper. I feel this would really tie all the tutorials together.

This issue is related to openjournals/joss-reviews#7291
