### 2024-10-30 mxwalbert/cohesivm#5 @ericfell
[JOSS Review] - dependencies section in README
Dear authors,

Can you also provide a **Dependencies** section in your root `README.md` listing the core dependencies? I find it most useful on a project `README` to quickly see which version(s) of Python is required.

This issue is related to openjournals/joss-reviews#7291
