### 2024-10-30 mxwalbert/cohesivm#6 @ericfell
[JOSS Review] - Interface motivation, docs/JOSS paper
Dear authors,

The `Interface` module inherently requires an array of electrode contacts (e.g. the MA8X8 array which you've open sourced). I think that `COHESIVM`'s requirement that a device must specifically be an electrode contact array configuration should be **explicitly** stated early on in both the root `README` and the JOSS paper, as this isn't made clear.

The `README`'s listed **Key Features** of _contacting interfaces (`Interface`)... are abstracted into interchangeable units_ is not readily apparent, given the above requirement.

Furthermore, the only interface provided (the in-house MA8X8) appears very specific to your lab. Without more provided examples of interfaces, realistically how applicable is COHESIVM to other groups? Or should this package be refactored specifically as a use case for your specific interface?

I understand my issue seems vague... I totally agree that a modular architecture built of abstract interchangeable units enables flexibility in designing a new system, but the description of the architecture must involve concrete examples to fully motivate its use. I'll add that there is overlap in my comments here and those of enricgrau in the JOSS review thread.

This issue is related to openjournals/joss-reviews#7291
