### 2024-11-17 c42336d
docs: fix incorrect dev instructions in README (#1)
