### 2024-11-17 ffd36f2
docs: added a "getting involved" (contact & contributing) section. added an admonition in tutorials to promote contributions. (#2)
