+++ docs/source/conf.py
@@ -25,7 +25,8 @@ extensions = [
     'sphinx.ext.napoleon',
     'sphinx.ext.viewcode',
     'sphinx_rtd_theme',
-    'sphinx.ext.mathjax'
+    'sphinx.ext.mathjax',
+    'sphinx_design'
 ]
 
 intersphinx_mapping = {
