+++ docs/source/readme/examples.md
@@ -1,4 +1,9 @@
-# Examples <a name="examples"></a>
+<a name="examples"></a>
+# Examples
+
+````{note}
+For a practical example, read the {doc}`Real-World Example</tutorials/real-world_example>`.
+````
 
 <!--- examples/experiment.md -->
 
