+++ docs/source/readme/getting_started.md
@@ -1,7 +1,11 @@
-# Getting Started <a name="getting-started"></a>
+<a name="getting-started"></a>
+# Getting Started
+
+<!--- getting_started/dependencies.md -->
 
 <!--- getting_started/installation.md -->
 
 <!--- getting_started/configuration.md -->
 
+<a name="basic-usage"></a>
 <!--- getting_started/basic_usage.md -->
