+++ docs/source/readme/contributing.md
@@ -1,3 +1,4 @@
-## Contributing <a name="contributing"></a>
+<a name="contributing"></a>
+# Contributing
 
 The contributing guidelines can be found [here](CONTRIBUTING.md).
