+++ docs/requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile '.\requirements.in'
+#    pip-compile --output-file=requirements.txt requirements.in
 #
 alabaster==0.7.16
     # via sphinx
@@ -11,7 +11,7 @@ asttokens==2.4.1
 babel==2.16.0
     # via sphinx
 bqplot==0.12.43
-    # via -r .\requirements.in
+    # via -r requirements.in
 certifi==2024.7.4
     # via requests
 charset-normalizer==3.3.2
@@ -41,7 +41,7 @@ executing==2.0.1
 fonttools==4.53.1
     # via matplotlib
 h5py==3.11.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 idna==3.7
     # via requests
 imagesize==1.4.1
@@ -71,7 +71,7 @@ markdown-it-py==3.0.0
 markupsafe==2.1.5
     # via jinja2
 matplotlib==3.9.1.post1
-    # via -r .\requirements.in
+    # via -r requirements.in
 matplotlib-inline==0.1.7
     # via ipython
 mdit-py-plugins==0.4.1
@@ -79,10 +79,10 @@ mdit-py-plugins==0.4.1
 mdurl==0.1.2
     # via markdown-it-py
 myst-parser==3.0.1
-    # via -r .\requirements.in
+    # via -r requirements.in
 numpy==1.26.4
     # via
-    #   -r .\requirements.in
+    #   -r requirements.in
     #   bqplot
     #   contourpy
     #   h5py
@@ -109,7 +109,7 @@ pygments==2.18.0
 pyparsing==3.1.2
     # via matplotlib
 pyserial==3.5
-    # via -r .\requirements.in
+    # via -r requirements.in
 python-dateutil==2.9.0.post0
     # via
     #   matplotlib
@@ -117,7 +117,7 @@ python-dateutil==2.9.0.post0
 pytz==2024.1
     # via pandas
 pyvisa==1.14.1
-    # via -r .\requirements.in
+    # via -r requirements.in
 pyyaml==6.0.2
     # via myst-parser
 requests==2.32.3
@@ -130,12 +130,15 @@ snowballstemmer==2.2.0
     # via sphinx
 sphinx==7.3.7
     # via
-    #   -r .\requirements.in
+    #   -r requirements.in
     #   myst-parser
+    #   sphinx-design
     #   sphinx-rtd-theme
     #   sphinxcontrib-jquery
+sphinx-design==0.6.1
+    # via -r requirements.in
 sphinx-rtd-theme==2.0.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 sphinxcontrib-applehelp==2.0.0
     # via sphinx
 sphinxcontrib-devhelp==2.0.0
@@ -155,7 +158,7 @@ stack-data==0.6.3
 tomli==2.0.1
     # via sphinx
 tqdm==4.66.5
-    # via -r .\requirements.in
+    # via -r requirements.in
 traitlets==5.14.3
     # via
     #   bqplot
@@ -179,7 +182,7 @@ wcwidth==0.2.13
 widgetsnbextension==4.0.11
     # via ipywidgets
 xtralien==2.11.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 zipp==3.20.0
     # via
     #   importlib-metadata

[…]
