### 2024-11-17 15a70a4
docs: added an example in the README's configuration section (#3)
