### 2024-11-17 187ea03
docs: added a real-world example in the tutorials to give a high-level description of using COHESIVM. (#4)
