### 2024-11-17 f2e0dae
docs: added a dependencies section. added admonitions in the basic usage example and the GUI description. (#5)
