### 2024-11-18 fe51301
docs/paper: linked the tutorials for component implementation and MA8X8 interface description in the README introduction. made the key features and advantages in the JOSS paper more explicit. referenced the real-world example to give a short use-case description in the JOSS paper. added author contributions to JOSS paper. (#6)
