### 2024-11-18 mxwalbert/cohesivm#2 @mxwalbert
Dear @ericfell,
we added a "Getting Involved" section in the documenation and put the contributiong guidelines and contact information there. Furthermore, we added invitations for contributions in the implementation tutorials in the readthedocs documentation.
