### 2024-11-18 mxwalbert/cohesivm#3 @mxwalbert
Dear @ericfell,
thank you for this suggestion. We added an example in the README's configuration section.
Please confirm if this gives a clearer picture of the configuration.
