### 2024-11-18 mxwalbert/cohesivm#4 @mxwalbert
Dear @ericfell,
thank you for this suggestion. We added a high-level description in the form of a real-world example, linked it in the README, and summarized it in the JOSS paper: https://github.com/mxwalbert/cohesivm/commit/fe513013142303ca911340f9b52389a0acd19a29
As you predicted, this really ties everything together and makes it more clear how and why one would use the package.
Please let us know if this turned out as expected.
