### 2024-11-18 mxwalbert/cohesivm#5 @mxwalbert
Dear @ericfell,
we added a dependencies section and included information about the available extras. We also added notes in the GUI descriptions, referring to the required ``gui`` extra.
