### 2024-11-18 mxwalbert/cohesivm#6 @mxwalbert
Dear @ericfell,

we deliberately designed the package to be reusable and I think refactoring it to a specific use case would make it less applicable to other groups.

Regarding the ``Interface`` component, I still consider it _interchangeable units_ since this can practically be anything as long as it somehow changes the contact position (which is something you would need if you want to measure combinatorial samples). Do you think it would help to implement a manual contact interface? This ``Interface`` could pause the ``Experiment`` to allow the user to physically change the contact position (e.g., by rewiring) before running the next ``Measurement``.

We added links for the implementation tutorials and for the MA8X8 description in the README, to make them more visible.

We also addressed the comments of @enricgrau regarding the JOSS paper in this issue:
- formulated explicit key features and advantages
- described a practical example for clarifying the usage of the package
- included author contributions

Please let us know if this contributes to resolving your issues and how we should proceed.
