### 2024-11-19 #7291 @RMeli
Hi everyone, how are things progressing? Could you please give me a quick update (just a few lines)? Thank you.

@enricgrau are you concerns being addressed? The number of commits should be taken with a pinch of salt, since it is somewhat arbitrary. The scholarly effort can be judged on the code itself, which in this case is ~4000 LOC (Python).
