### 2024-11-20 #7291 @mxwalbert
Hi @RMeli,
the proposed changes were just recently implemented by us. We tried to address all raised issues and commited the changes in the ``1-joss-review`` branch.
