### 2024-11-21 mxwalbert/cohesivm#3 @ericfell
Yes it does, thank you for adding this example.
I'll close this issue
