### 2024-11-22 #7291 @enricgrau
The manuscript is somewhat changed and slightly improved, thus it is a bit more clear. I would insist, however, that the claim that it has been used before citing an unpublished and unavailable manuscript is no good. Unless @RMeli thinks this is fine, I cannot accept this. If @mxwalbert wants this claim, they need to upload a pre-print to arXiv, Research Gate, or similar.
The repo itself looks great. I only created a minute issue.
