### 2024-11-23 #7291 @mxwalbert
Dear @enricgrau and @RMeli,
we can remove the unpublished work. It was only put there for future reference but since we now have the use-case description and linked tutorial, it's not needed anymore.
