### 2024-11-23 baf622a
removed future reference in JOSS paper (#6)
