+++ paper/paper.bib
@@ -32,12 +32,6 @@
   doi={10.1039/D3MA00136A}
 }
 
-@unpublished{wolf2024unpublished,
-  author={Wolf, Maximilian and Madsen, Georg KH and Dimopoulos, Theodoros},
-  title={Bayesian Optimization of Spray Parameters for the Deposition of {Ga}_2{O}_3-{Cu}_2{O} Semiconductor Heterojunctions},
-  year={unpublished}
-}
-
 @article{pernstich2012instrument,
   title={Instrument Control (iC)--an open-source software to automate test equipment},
   author={Pernstich, KP},
