### 2024-11-23 mxwalbert/cohesivm#6 @mxwalbert
Dear @enricgrau,
please confirm if you are fine with the paper now and close the issue.
