### 2024-12-20 #7291 @RMeli
Hi @mxwalbert, sorry for the delay but this time of the year is super busy. I had a look at the paper and I think it's OK, but I feel you should remove the whole
> So far, COHESIVM has been used for the investigation of oxide semiconductor heterojunctions where
it enables to quickly screen a matrix of 8&nbsp;×&nbsp;8 pixels on a single substrate
(25&nbsp;mm × 25&nbsp;mm)

sentence, or at least re-phrase it in a way that says how this application is described in the documentation. *The link to the documentation is broken, and needs to be fixed.*

(Ideally, you could put the previously referenced work on a pre-print server and cite it.)
