### 2025-01-03 72a916a
re-phrasing usage description in paper.md
