### 2025-01-03 #7291 @mxwalbert
Dear @RMeli, no worries I was very busy as well! I re-phrased the mentioned sentence to make it less reference-dependent. To resolve the broken link, I merged the ``1-joss-review`` to the main branch. (I had to clear my browser cache to be able to load the linked page.)
