### 2025-01-13 #7291 @RMeli
@mxwalbert can you please proceed with the following?
* [ ] Make a release of the software with the latest changes from the review and post the version number here.
* [ ] Archive the release on Zenodo/figshare/etc and post the DOI here

Thank you.
