### 2025-01-14 #7291 @mxwalbert
@RMeli

- v1.0.0-joss
- https://doi.org/10.5281/zenodo.14643235 (DOI maybe takes some time to be active: https://zenodo.org/records/14643235)
