### 2025-01-14 #7291 @RMeli
Thanks @mxwalbert. However, you should do a proper release. `v1.0.0-joss` don't follow semantic versioning (and that's fine, but it is what you want to do?), but for example this new version is not reflected in `pyproject.toml`, which is stuck at `1.0.0`.
