### 2025-01-14 #7291 @mxwalbert
Thanks for the quick response @RMeli! Since there were no modifications in the code but only in the documentation, I kept the version semantically at 1.0.0. This should also indicate that it is the same as the version on PyPI. So for the installed package defined in ``pyproject.toml``, there is practically no difference. Please let me know if this is fine for you, otherwise I can do a v1.0.1.
