### 2025-01-14 #7291 @RMeli
Thanks for the additional context @mxwalbert, I missed that some of the changes were made before the release. I discussed this internally with other editors, and since the documentation is part of the codebase, we think it would be better to have a proper `1.0.1` release if it is not too much to ask. Would that be OK with you?
