### 2025-01-27 #7291 @mxwalbert
Dear @RMeli, sorry for the delay. I will do the ``1.0.1`` release now and send you the updated Zenodo link.
