### 2025-01-27 #7291 @mxwalbert
@RMeli
- ``v1.0.1``
- [https://doi.org/10.5281/zenodo.14747442](https://doi.org/10.5281/zenodo.14747442)
