### 2025-02-03 #7291 @RMeli
@mxwalbert one last nitpick, then it is good to go: can you please remove the `v1.0.1` from the title of the Zenodo archive? The title of the archive should match the title of the paper. Thank you.
