### 2025-02-03 #7291 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
