### 2025-02-05 #7291 @Kevin-Mattheus-Moerman
@mxwalbert as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:
